Add test to ensure exit at map edge
