+++ game.js
@@ -1,6 +1,6 @@
-import{GRID_W,GRID_H,PASSIVE_MANA,START_MANA,START_HP,CHEST_MANA,COSTS,TRAP_RANGE,TRAP_DMG,RUNE_RADIUS,FIRE_DMG,FIRE_RADIUS,SAB_EXP_DMG,SAB_EXP_RADIUS,SPIKE_DMG,PLACE_RADIUS,PLACE_ZOOM,ARROW_AMMO,BURN_TURNS,BURN_DMG,RUNE_SLOW_TURNS,DASH_CD,DASH_COST,DENSITY_TILE_WEIGHT,DENSITY_NEIGHBOR_WEIGHT,PATIENCE_PROB,PATROL_RADIUS,ENEMY,baseSpawnCooldown,baseSpawnCount,ENEMY_CAP,CHESTS_PER_RUN,SPAWN_MIN_RADIUS,COLORS}from './constants.js';
+import{GRID_W,GRID_H,VIEW_W,PASSIVE_MANA,START_MANA,START_HP,CHEST_MANA,COSTS,TRAP_RANGE,TRAP_DMG,RUNE_RADIUS,FIRE_DMG,FIRE_RADIUS,SAB_EXP_DMG,SAB_EXP_RADIUS,SPIKE_DMG,PLACE_RADIUS,PLACE_ZOOM,ARROW_AMMO,BURN_TURNS,BURN_DMG,RUNE_SLOW_TURNS,DASH_CD,DASH_COST,DENSITY_TILE_WEIGHT,DENSITY_NEIGHBOR_WEIGHT,PATIENCE_PROB,PATROL_RADIUS,ENEMY,baseSpawnCooldown,baseSpawnCount,ENEMY_CAP,CHESTS_PER_RUN,SPAWN_MIN_RADIUS,COLORS}from './constants.js';
 import './ui.js';
-import './map.js';
+import { buildMap } from './map.js';
 import './enemies.js';
 
 (() => {
@@ -22,19 +22,14 @@ function renderLegend(pairs){const cont=document.getElementById('legend-items');
 function setDashArmed(armed,cd){btnDash.disabled=cd>0;btnDash.textContent=armed?(cd>0?`Dash (cd: ${cd})`:'Dash (ready)'):'Arm Dash';}
 function updateHUD(st=state){hud.hp.textContent=st.hp|0;hud.turn.textContent=st.turn|0;hud.enemy.textContent=st.enemies.length|0;hud.spawn.textContent=st.nextSpawn|0;hud.dash.textContent=st.dashCD>0?st.dashCD:'Ready';setDashArmed(st.dashArmed,st.dashCD);setActiveTrap(st.selectedTool);updateMana(st.mana);}
 
+function updateCamera(){state.cameraX=Math.max(0,Math.min(state.player.x-Math.floor(VIEW_W/2),GRID_W-VIEW_W));state.cameraY=0;}
+
 const LEGEND_DATA=[['Wall',COLORS.wall],['Floor',COLORS.floor],['Entrance',COLORS.start],['Exit',COLORS.exit],['Spawner',COLORS.spawner],['Chest',COLORS.chest],['Goblin',COLORS.enemyGoblin],['Archer',COLORS.enemyArcher],['Wraith',COLORS.enemyWraith],['Brute',COLORS.enemyBrute],['Saboteur',COLORS.enemySaboteur],['Hunter',COLORS.enemyHunter]];
 
-function generateMaze(width,height){const w=(width%2===0)?width-1:width,h=(height%2===0)?height-1:height;const grid=Array.from({length:height},()=>Array(width).fill(1));function inBoundsCarve(x,y){return x>0&&x<w-1&&y>0&&y<h-1}const stack=[];let cx=1,cy=1;grid[cy][cx]=0;stack.push([cx,cy]);function neighbors(x,y){return[[x+2,y],[x-2,y],[x,y+2],[x,y-2]].filter(([nx,ny])=>inBoundsCarve(nx,ny)&&grid[ny][nx]===1)}while(stack.length){const [x,y]=stack[stack.length-1];const nbs=neighbors(x,y);if(!nbs.length){stack.pop();continue}const [nx,ny]=nbs[(Math.random()*nbs.length)|0];grid[ny][nx]=0;grid[y+(ny-y)/2][x+(nx-x)/2]=0;stack.push([nx,ny])}for(let y=1;y<height-1;y++){for(let x=1;x<width-1;x++){if(grid[y][x]===0&&Math.random()<.22){if(grid[y][x+1]===1)grid[y][x+1]=0;if(grid[y+1][x]===1)grid[y+1][x]=0}}}return grid}
-function carveRect(grid,x,y,w,h){for(let j=0;j<h;j++)for(let i=0;i<w;i++){const gx=x+i,gy=y+j;if(gx>0&&gy>0&&gx<GRID_W-1&&gy<GRID_H-1)grid[gy][gx]=0}}
-function carveGuidedPath(grid,start,exit){let x=start.x,y=start.y;grid[y][x]=0;if(y+1<GRID_H)grid[y+1][x]=0;const steps=GRID_W*2;let dirY=0;for(let s=0;s<steps&&x<exit.x-1;s++){const r=Math.random();if(r<.70){x=Math.min(GRID_W-2,x+1)}else{if(dirY===0)dirY=Math.random()<.5?-1:1;else if(Math.random()<.4)dirY=0;y=Math.max(1,Math.min(GRID_H-3,y+dirY))}grid[y][x]=0;grid[y+1][x]=0;if(Math.random()<.30&&x<exit.x-2){grid[y][x+1]=1;grid[y+1][x+1]=1;const off=(Math.random()<.5?-1:1);const yy=Math.max(1,Math.min(GRID_H-3,y+off));grid[yy][x]=0;grid[yy+1][x]=0;grid[yy][x+2]=0;grid[yy+1][x+2]=0;x=x+2;y=yy}}grid[exit.y][exit.x]=0;if(exit.x-1>=0)grid[exit.y][exit.x-1]=0}
-function addRoomsAndConnectors(grid){const roomCount=3+((Math.random()*2)|0);for(let r=0;r<roomCount;r++){const rw=3+((Math.random()*3)|0),rh=3+((Math.random()*2)|0),rx=2+((Math.random()*(GRID_W-rw-4))|0),ry=2+((Math.random()*(GRID_H-rh-4))|0);carveRect(grid,rx,ry,rw,rh)}for(let y=2;y<GRID_H-2;y++){for(let x=2;x<GRID_W-2;x++){if(grid[y][x]!==1)continue;const horiz=(grid[y][x-1]===0&&grid[y][x+1]===0&&grid[y-1][x]===1&&grid[y+1][x]===1);const vert=(grid[y-1][x]===0&&grid[y+1][x]===0&&grid[y][x-1]===1&&grid[y][x+1]===1);if((horiz||vert)&&Math.random()<.08)grid[y][x]=0}}}
-function randomFloor(grid){for(let tries=0;tries<6000;tries++){const x=(Math.random()*GRID_W)|0,y=(Math.random()*GRID_H)|0;if(grid[y][x]===0)return{x,y}}return{x:1,y:1}}
-function dist1(a,b){return Math.abs(a.x-b.x)+Math.abs(a.y-b.y)}
-function pathExists(grid,start,goal){const q=[start];const seen=new Set([start.x+','+start.y]);const dirs=[[1,0],[-1,0],[0,1],[0,-1]];while(q.length){const cur=q.shift();if(cur.x===goal.x&&cur.y===goal.y)return true;for(let i=0;i<4;i++){const nx=cur.x+dirs[i][0],ny=cur.y+dirs[i][1];if(nx<0||ny<0||nx>=GRID_W||ny>=GRID_H)continue;if(grid[ny][nx]===1)continue;const key=nx+','+ny;if(!seen.has(key)){seen.add(key);q.push({x:nx,y:ny})}}}return false}
-function buildMap(){let grid,start,exit,attempts=0;do{grid=generateMaze(GRID_W,GRID_H);start={x:0,y:(GRID_H/2)|0};exit={x:GRID_W-1,y:(GRID_H/2)|0};grid[start.y][start.x]=0;if(start.x+1<GRID_W)grid[start.y][start.x+1]=0;carveGuidedPath(grid,start,exit);addRoomsAndConnectors(grid);attempts++;if(attempts>50)break}while(!pathExists(grid,start,exit));const spawners=[];const edgeOptions=[];for(let y=0;y<GRID_H;y++){if(grid[y][0]===0)edgeOptions.push({x:0,y});if(grid[y][GRID_W-1]===0)edgeOptions.push({x:GRID_W-1,y})}for(let x=0;x<GRID_W;x++){if(grid[0][x]===0)edgeOptions.push({x,y:0});if(grid[GRID_H-1][x]===0)edgeOptions.push({x,y:GRID_H-1})}edgeOptions.sort(()=>Math.random()-.5);for(const p of edgeOptions){if(dist1(p,start)<SPAWN_MIN_RADIUS)continue;if(!((p.x===start.x&&p.y===start.y)||(p.x===exit.x&&p.y===exit.y)))spawners.push(p);if(spawners.length>=3)break}while(spawners.length<3){const p=randomFloor(grid);if(dist1(p,start)<SPAWN_MIN_RADIUS)continue;if((p.x===start.x&&p.y===start.y)||(p.x===exit.x&&p.y===exit.y))continue;spawners.push(p)}if(!spawners.some(s=>s.x>=Math.floor(GRID_W*.6))){for(let tries=0;tries<200;tries++){const x=Math.floor(GRID_W*.7)+((Math.random()*Math.floor(GRID_W*.3))|0);const y=(Math.random()*GRID_H)|0;const p={x,y};if(x>=0&&x<GRID_W&&y>=0&&y<GRID_H&&grid[y][x]===0&&dist1(p,start)>=SPAWN_MIN_RADIUS){spawners[0]=p;break}}}const chests=[];for(let i=0;i<CHESTS_PER_RUN;i++){const p=randomFloor(grid);if((p.x===start.x&&p.y===start.y)||(p.x===exit.x&&p.y===exit.y)){i--;continue}if(spawners.some(s=>s.x===p.x&&s.y===p.y)){i--;continue}chests.push({x:p.x,y:p.y,opened:false})}return{grid,start,exit,spawners,chests}}
+
 
 function inBounds(x,y){return x>=0&&x<GRID_W&&y>=0&&y<GRID_H}function samePos(a,b){return a.x===b.x&&a.y===b.y}
-function tileToScreen(x,y){return{sx:x*tileSize,sy:y*tileSize}}function isWall(x,y){return state.map.grid[y][x]===1}
+function tileToScreen(x,y,useCam=true){const cx=useCam?state.cameraX:0,cy=useCam?state.cameraY:0;return{sx:(x-cx)*tileSize,sy:(y-cy)*tileSize}}function isWall(x,y){return state.map.grid[y][x]===1}
 function isSpawner(x,y){return state.map.spawners.some(s=>s.x===x&&s.y===y)}function isChest(x,y){return state.map.chests.some(c=>!c.opened&&c.x===x&&c.y===y)}
 function isStart(x,y){return state.map.start.x===x&&state.map.start.y===y}function isExit(x,y){return state.map.exit.x===x&&state.map.exit.y===y}
 function logMsg(m){const p=document.createElement('p');p.textContent=m;hud.log.appendChild(p);hud.log.scrollTop=hud.log.scrollHeight}
@@ -42,10 +37,10 @@ function clearLog(){hud.log.innerHTML=""}function rndShuffle(a){for(let i=a.leng
 
 function lineOfSightRowCol(a,b){if(a.x===b.x){const x=a.x;const y1=Math.min(a.y,b.y)+1,y2=Math.max(a.y,b.y);for(let y=y1;y<y2;y++)if(isWall(x,y))return false;return true}else if(a.y===b.y){const y=a.y;const x1=Math.min(a.x,b.x)+1,x2=Math.max(a.x,b.x);for(let x=x1;x<x2;x++)if(isWall(x,y))return false;return true}return false}
 function clearShotToPlayer(from,ignore=null){if(!lineOfSightRowCol(from,state.player))return false;if(from.x===state.player.x){const x=from.x,y1=Math.min(from.y,state.player.y)+1,y2=Math.max(from.y,state.player.y);for(let y=y1;y<y2;y++)if(state.enemies.some(e=>e!==ignore&&e.x===x&&e.y===y))return false}else if(from.y===state.player.y){const y=from.y,x1=Math.min(from.x,state.player.x)+1,x2=Math.max(from.x,state.player.x);for(let x=x1;x<x2;x++)if(state.enemies.some(e=>e!==ignore&&e.x===x&&e.y===y))return false}return true}
-function ensureOffscreen(){if(!terrainCanvas){terrainCanvas=document.createElement('canvas');terrainCtx=terrainCanvas.getContext('2d',{alpha:false})||terrainCanvas.getContext('2d')}const rect=canvas.getBoundingClientRect();terrainCanvas.width=Math.floor(rect.width);terrainCanvas.height=Math.floor(rect.height);terrainCtx.setTransform(1,0,0,1,0,0);terrainValid=false}
-function drawWallTileTo(tctx,x,y){const {sx,sy}=tileToScreen(x,y);tctx.fillStyle=COLORS.wall;tctx.fillRect(sx,sy,tileSize,tileSize);tctx.strokeStyle=COLORS.wallEdge;tctx.lineWidth=Math.max(1,tileSize*.06);tctx.strokeRect(sx+.5,sy+.5,tileSize-1,tileSize-1);tctx.save();tctx.beginPath();tctx.rect(sx+1,sy+1,tileSize-2,tileSize-2);tctx.clip();tctx.strokeStyle='rgba(255,255,255,.06)';tctx.lineWidth=Math.max(1,tileSize*.05);const step=Math.max(4,tileSize/4);for(let k=-tileSize;k<tileSize*2;k+=step){tctx.beginPath();tctx.moveTo(sx+k,sy);tctx.lineTo(sx,sy+k);tctx.stroke()}tctx.restore()}
-function drawTerrainAll(){const rect=canvas.getBoundingClientRect();terrainCtx.clearRect(0,0,rect.width,rect.height);for(let y=0;y<GRID_H;y++)for(let x=0;x<GRID_W;x++){if(isWall(x,y))drawWallTileTo(terrainCtx,x,y);else{const {sx,sy}=tileToScreen(x,y);terrainCtx.fillStyle=COLORS.floor;terrainCtx.fillRect(sx+tilePad,sy+tilePad,tileSize-tilePad*2,tileSize-tilePad*2)}}drawOutlineRectTo(terrainCtx,state.map.start.x,state.map.start.y,COLORS.start,.35);drawOutlineRectTo(terrainCtx,state.map.exit.x,state.map.exit.y,COLORS.exit,.35);for(const s of state.map.spawners)drawOutlineRectTo(terrainCtx,s.x,s.y,COLORS.spawner,.35);for(const c of state.map.chests)if(!c.opened){drawOutlineRectTo(terrainCtx,c.x,c.y,COLORS.chest,.45);const {sx,sy}=tileToScreen(c.x,c.y);const s2=Math.max(4,tileSize*.35);terrainCtx.fillStyle=COLORS.chest;terrainCtx.fillRect(sx+(tileSize-s2)/2,sy+(tileSize-s2)/2,s2,s2)}terrainValid=true}
-function drawOutlineRectTo(tctx,x,y,color,alpha=.28){const {sx,sy}=tileToScreen(x,y);tctx.save();tctx.globalAlpha=alpha;tctx.strokeStyle=color;tctx.setLineDash([4,3]);tctx.lineWidth=Math.max(1,tileSize*.06);tctx.strokeRect(sx+tilePad,sy+tilePad,tileSize-tilePad*2,tileSize-tilePad*2);tctx.restore()}
+function ensureOffscreen(){if(!terrainCanvas){terrainCanvas=document.createElement('canvas');terrainCtx=terrainCanvas.getContext('2d',{alpha:false})||terrainCanvas.getContext('2d')}const w=Math.floor(tileSize*GRID_W),h=Math.floor(tileSize*GRID_H);if(terrainCanvas.width!==w||terrainCanvas.height!==h){terrainCanvas.width=w;terrainCanvas.height=h;terrainCtx.setTransform(1,0,0,1,0,0);terrainValid=false}}
+function drawWallTileTo(tctx,x,y){const {sx,sy}=tileToScreen(x,y,false);tctx.fillStyle=COLORS.wall;tctx.fillRect(sx,sy,tileSize,tileSize);tctx.strokeStyle=COLORS.wallEdge;tctx.lineWidth=Math.max(1,tileSize*.06);tctx.strokeRect(sx+.5,sy+.5,tileSize-1,tileSize-1);tctx.save();tctx.beginPath();tctx.rect(sx+1,sy+1,tileSize-2,tileSize-2);tctx.clip();tctx.strokeStyle='rgba(255,255,255,.06)';tctx.lineWidth=Math.max(1,tileSize*.05);const step=Math.max(4,tileSize/4);for(let k=-tileSize;k<tileSize*2;k+=step){tctx.beginPath();tctx.moveTo(sx+k,sy);tctx.lineTo(sx,sy+k);tctx.stroke()}tctx.restore()}
+function drawTerrainAll(){terrainCtx.clearRect(0,0,terrainCanvas.width,terrainCanvas.height);for(let y=0;y<GRID_H;y++)for(let x=0;x<GRID_W;x++){if(isWall(x,y))drawWallTileTo(terrainCtx,x,y);else{const {sx,sy}=tileToScreen(x,y,false);terrainCtx.fillStyle=COLORS.floor;terrainCtx.fillRect(sx+tilePad,sy+tilePad,tileSize-tilePad*2,tileSize-tilePad*2)}}drawOutlineRectTo(terrainCtx,state.map.start.x,state.map.start.y,COLORS.start,.35,false);drawOutlineRectTo(terrainCtx,state.map.exit.x,state.map.exit.y,COLORS.exit,.35,false);for(const s of state.map.spawners)drawOutlineRectTo(terrainCtx,s.x,s.y,COLORS.spawner,.35,false);for(const c of state.map.chests)if(!c.opened){drawOutlineRectTo(terrainCtx,c.x,c.y,COLORS.chest,.45,false);const {sx,sy}=tileToScreen(c.x,c.y,false);const s2=Math.max(4,tileSize*.35);terrainCtx.fillStyle=COLORS.chest;terrainCtx.fillRect(sx+(tileSize-s2)/2,sy+(tileSize-s2)/2,s2,s2)}terrainValid=true}
+function drawOutlineRectTo(tctx,x,y,color,alpha=.28,useCam=true){const {sx,sy}=tileToScreen(x,y,useCam);tctx.save();tctx.globalAlpha=alpha;tctx.strokeStyle=color;tctx.setLineDash([4,3]);tctx.lineWidth=Math.max(1,tileSize*.06);tctx.strokeRect(sx+tilePad,sy+tilePad,tileSize-tilePad*2,tileSize-tilePad*2);tctx.restore()}
 function drawOutlineRect(x,y,c,a){drawOutlineRectTo(ctx,x,y,c,a)}
 function drawHPBar(x,y,ratio){const {sx,sy}=tileToScreen(x,y);const w=tileSize-tilePad*2,h=Math.max(3,tileSize*.09);const bx=sx+tilePad,by=sy+tilePad*.7;ctx.fillStyle='rgba(0,0,0,.5)';ctx.fillRect(bx,by,w,h);ctx.fillStyle=ratio>.5?'#22c55e':(ratio>.25?'#f59e0b':'#ef4444');ctx.fillRect(bx,by,w*ratio,h)}
 function drawTrapMeter(t){const {sx,sy}=tileToScreen(t.x,t.y);const ratio=(t.ammo===undefined?1:t.ammo/ARROW_AMMO);const w=tileSize-tilePad*2,h=Math.max(3,tileSize*.1);const bx=sx+tilePad,by=sy+tileSize-h-tilePad;ctx.fillStyle='rgba(0,0,0,.5)';ctx.fillRect(bx,by,w,h);ctx.fillStyle=ratio>.5?'#22c55e':(ratio>.25?'#f59e0b':'#ef4444');ctx.fillRect(bx,by,w*ratio,h)}
@@ -134,11 +129,11 @@ function drawEffects(){
 
 function drawPlayer(){const {sx,sy}=tileToScreen(state.player.x,state.player.y);const bob=Math.sin(animT/200)*tileSize*.1;ctx.save();ctx.translate(sx+tileSize/2,sy+tileSize/2+bob);const r=tileSize/2-tilePad;ctx.fillStyle=COLORS.player;ctx.beginPath();ctx.arc(0,0,r,0,Math.PI*2);ctx.fill();ctx.fillStyle='#000';const eyeOffset=tileSize*.15,eyeR=tileSize*.07;ctx.beginPath();ctx.arc(-eyeOffset,-eyeOffset,eyeR,0,Math.PI*2);ctx.arc(eyeOffset,-eyeOffset,eyeR,0,Math.PI*2);ctx.fill();ctx.restore()}
 function drawEnemy(e){const {sx,sy}=tileToScreen(e.x,e.y);const bob=Math.sin(animT/200+(e.x+e.y))*tileSize*.1;ctx.save();ctx.translate(sx+tileSize/2,sy+tileSize/2+bob);const r=tileSize/2-tilePad;let col;switch(e.kind){case'goblin':col=COLORS.enemyGoblin;break;case'archer':col=COLORS.enemyArcher;break;case'wraith':col=COLORS.enemyWraith;break;case'brute':col=COLORS.enemyBrute;break;case'saboteur':col=COLORS.enemySaboteur;break;case'hunter':col=COLORS.enemyHunter;break;default:col=COLORS.enemyWraith;}ctx.fillStyle=col;ctx.beginPath();ctx.arc(0,0,r,0,Math.PI*2);ctx.fill();ctx.fillStyle='#000';const eyeR=tileSize*.07;ctx.beginPath();ctx.arc(-r*.3,-r*.2,eyeR,0,Math.PI*2);ctx.arc(r*.3,-r*.2,eyeR,0,Math.PI*2);ctx.fill();if(e.kind==='archer'){ctx.strokeStyle='#000';ctx.lineWidth=Math.max(2,tileSize*.07);ctx.beginPath();ctx.moveTo(-r*.6,0);ctx.lineTo(r*.6,0);ctx.stroke()}if(e.kind==='wraith'){ctx.fillStyle='rgba(0,0,0,.3)';ctx.beginPath();ctx.moveTo(-r,r*.2);ctx.lineTo(0,r);ctx.lineTo(r,r*.2);ctx.closePath();ctx.fill()}ctx.restore();const maxhp=e.maxhp||(ENEMY[e.kind]?ENEMY[e.kind].hp:ENEMY.wraith.hp);drawHPBar(e.x,e.y,Math.max(0,e.hp)/maxhp)}
-function draw(){const rect=canvas.getBoundingClientRect();const dpr=window.devicePixelRatio||1;canvas.width=Math.floor(rect.width*dpr);canvas.height=Math.floor(rect.height*dpr);ctx.setTransform(dpr,0,0,dpr,0,0);tileSize=rect.width/GRID_W;tilePad=Math.max(1,Math.floor(tileSize*.03));ensureOffscreen();if(!terrainValid)drawTerrainAll();ctx.save();if(state.placeMode){const {sx:psx,sy:psy}=tileToScreen(state.player.x,state.player.y);ctx.translate(rect.width/2,rect.height/2);ctx.scale(PLACE_ZOOM,PLACE_ZOOM);ctx.translate(-psx-tileSize/2,-psy-tileSize/2);}ctx.drawImage(terrainCanvas,0,0,rect.width,rect.height);if(state.placeMode)highlightPlacementArea();const pulse=(Math.sin(animT/600)+1)/2;for(const s of state.map.spawners){const {sx,sy}=tileToScreen(s.x,s.y);ctx.save();ctx.globalAlpha=.35+.35*pulse;ctx.strokeStyle=COLORS.spawner;ctx.lineWidth=Math.max(2,tileSize*.1);ctx.beginPath();ctx.arc(sx+tileSize/2,sy+tileSize/2,tileSize*.42+tileSize*.08*pulse,0,Math.PI*2);ctx.stroke();ctx.restore()}if(state.placeMode){for(const t of state.towers){if(t.type==='arrow')outlineRangeTiles(t.x,t.y,TRAP_RANGE,COLORS.arrow);if(t.type==='rune')outlineRangeTiles(t.x,t.y,RUNE_RADIUS,COLORS.rune);if(t.type==='fire')outlineRangeTiles(t.x,t.y,FIRE_RADIUS,COLORS.fire);if(t.type==='spike')drawOutlineRect(t.x,t.y,COLORS.spike,.25)}}for(const t of state.towers){drawTrapIcon(t);if(t.ammo!==undefined)drawTrapMeter(t)}for(const e of state.enemies)drawEnemy(e);drawPlayer();drawEffects();ctx.restore();if(state.won||state.lost){ctx.save();ctx.fillStyle='rgba(0,0,0,.55)';ctx.fillRect(0,0,rect.width,rect.height);ctx.fillStyle=state.won?'#7dff9d':'#ff6b6b';ctx.font='bold 26px system-ui';ctx.textAlign='center';ctx.textBaseline='middle';ctx.fillText(state.won?'YOU ESCAPED!':'DEFEATED',rect.width/2,rect.height/2-10);ctx.fillStyle='#e8ecff';ctx.font='16px system-ui';ctx.fillText('Tap "New Game" to try again',rect.width/2,rect.height/2+18);ctx.restore()}}window.addEventListener('resize',()=>{ensureOffscreen();terrainValid=false});
+function draw(){const rect=canvas.getBoundingClientRect();const dpr=window.devicePixelRatio||1;canvas.width=Math.floor(rect.width*dpr);canvas.height=Math.floor(rect.height*dpr);ctx.setTransform(dpr,0,0,dpr,0,0);tileSize=rect.width/VIEW_W;tilePad=Math.max(1,Math.floor(tileSize*.03));updateCamera();ensureOffscreen();if(!terrainValid)drawTerrainAll();ctx.save();if(state.placeMode){const {sx:psx,sy:psy}=tileToScreen(state.player.x,state.player.y);ctx.translate(rect.width/2,rect.height/2);ctx.scale(PLACE_ZOOM,PLACE_ZOOM);ctx.translate(-psx-tileSize/2,-psy-tileSize/2);}ctx.drawImage(terrainCanvas,-state.cameraX*tileSize,-state.cameraY*tileSize);if(state.placeMode)highlightPlacementArea();const pulse=(Math.sin(animT/600)+1)/2;for(const s of state.map.spawners){const {sx,sy}=tileToScreen(s.x,s.y);ctx.save();ctx.globalAlpha=.35+.35*pulse;ctx.strokeStyle=COLORS.spawner;ctx.lineWidth=Math.max(2,tileSize*.1);ctx.beginPath();ctx.arc(sx+tileSize/2,sy+tileSize/2,tileSize*.42+tileSize*.08*pulse,0,Math.PI*2);ctx.stroke();ctx.restore()}if(state.placeMode){for(const t of state.towers){if(t.type==='arrow')outlineRangeTiles(t.x,t.y,TRAP_RANGE,COLORS.arrow);if(t.type==='rune')outlineRangeTiles(t.x,t.y,RUNE_RADIUS,COLORS.rune);if(t.type==='fire')outlineRangeTiles(t.x,t.y,FIRE_RADIUS,COLORS.fire);if(t.type==='spike')drawOutlineRect(t.x,t.y,COLORS.spike,.25)}}for(const t of state.towers){drawTrapIcon(t);if(t.ammo!==undefined)drawTrapMeter(t)}for(const e of state.enemies)drawEnemy(e);drawPlayer();drawEffects();ctx.restore();if(state.won||state.lost){ctx.save();ctx.fillStyle='rgba(0,0,0,.55)';ctx.fillRect(0,0,rect.width,rect.height);ctx.fillStyle=state.won?'#7dff9d':'#ff6b6b';ctx.font='bold 26px system-ui';ctx.textAlign='center';ctx.textBaseline='middle';ctx.fillText(state.won?'YOU ESCAPED!':'DEFEATED',rect.width/2,rect.height/2-10);ctx.fillStyle='#e8ecff';ctx.font='16px system-ui';ctx.fillText('Tap "New Game" to try again',rect.width/2,rect.height/2+18);ctx.restore()}}window.addEventListener('resize',()=>{ensureOffscreen();terrainValid=false});
 document.addEventListener('keydown',(e)=>{if(state.won||state.lost||state.placeMode)return;const key=e.key.toLowerCase();if(['arrowup','w'].includes(key))playerMove(0,-1,e.shiftKey);else if(['arrowdown','s'].includes(key))playerMove(0,1,e.shiftKey);else if(['arrowleft','a'].includes(key))playerMove(-1,0,e.shiftKey);else if(['arrowright','d'].includes(key))playerMove(1,0,e.shiftKey);else if(key==='q')toggleDashArm();else if(['1','2','3','4'].includes(e.key))setActiveTrap(TRAP_DEFS[Number(e.key)-1].id);});
 btnDash.addEventListener('click',()=>{if(!btnDash.disabled)toggleDashArm()});
 function toggleDashArm(){if(state.dashCD>0||state.mana<DASH_COST){logMsg(state.dashCD>0?`Dash on cooldown (${state.dashCD}).`:`Need ${DASH_COST} mana to dash.`);return}state.dashArmed=!state.dashArmed;setDashArmed(state.dashArmed,state.dashCD)}
-function canvasPosToTile(clientX,clientY){const rect=canvas.getBoundingClientRect();let x=clientX-rect.left,y=clientY-rect.top;if(state.placeMode){const {sx:psx,sy:psy}=tileToScreen(state.player.x,state.player.y);x=(x-rect.width/2)/PLACE_ZOOM+psx+tileSize/2;y=(y-rect.height/2)/PLACE_ZOOM+psy+tileSize/2}x=Math.floor(Math.max(0,Math.min(rect.width-1,x))/(rect.width/GRID_W));y=Math.floor(Math.max(0,Math.min(rect.height-1,y))/(rect.height/GRID_H));return{x,y}}
+function canvasPosToTile(clientX,clientY){const rect=canvas.getBoundingClientRect();let x=clientX-rect.left,y=clientY-rect.top;if(state.placeMode){const {sx:psx,sy:psy}=tileToScreen(state.player.x,state.player.y);x=(x-rect.width/2)/PLACE_ZOOM+psx+tileSize/2;y=(y-rect.height/2)/PLACE_ZOOM+psy+tileSize/2}x=Math.floor(Math.max(0,Math.min(rect.width-1,x))/tileSize)+state.cameraX;y=Math.floor(Math.max(0,Math.min(rect.height-1,y))/tileSize)+state.cameraY;return{x,y}}
 btnPlace.addEventListener('click',()=>{if(btnPlace.disabled)return;state.placeMode=!state.placeMode;if(!state.placeMode)state.hover=null;drawPlacementPreview();});
 canvas.addEventListener('click',(e)=>{if(!state.placeMode||state.won||state.lost)return;const {x,y}=canvasPosToTile(e.clientX,e.clientY);const check=isValidPlacement(x,y);if(!check.ok){logMsg(`Can't place: ${check.reason}`);return}tryPlace(x,y)},{passive:true});
 canvas.addEventListener('mousemove',(e)=>{if(!state.placeMode){drawPlacementPreview();return;}const {x,y}=canvasPosToTile(e.clientX,e.clientY);drawPlacementPreview(x,y);},{passive:true});
@@ -149,7 +144,7 @@ canvas.addEventListener('touchstart',e=>{const t=e.changedTouches[0];swipeStart=
 canvas.addEventListener('touchend',e=>{if(!swipeStart)return;const t=e.changedTouches[0];const dx=t.clientX-swipeStart.x,dy=t.clientY-swipeStart.y;if(Math.abs(dx)>Math.abs(dy)&&Math.abs(dx)>24)window.onMove(dx>0?'right':'left');else if(Math.abs(dy)>24)window.onMove(dy>0?'down':'up');swipeStart=null;},{passive:true});
 function rebuildFlow(){const INF=1e9;const dist=Array.from({length:GRID_H},()=>Array(GRID_W).fill(INF));const q=[];const start={x:state.player.x,y:state.player.y};dist[start.y][start.x]=0;q.push(start);const dirs=[[1,0],[-1,0],[0,1],[0,-1]];while(q.length){const cur=q.shift();const d=dist[cur.y][cur.x]+1;for(let i=0;i<4;i++){const nx=cur.x+dirs[i][0],ny=cur.y+dirs[i][1];if(!inBounds(nx,ny))continue;if(isWall(nx,ny))continue;if(d<dist[ny][nx]){dist[ny][nx]=d;q.push({x:nx,y:ny})}}}state.flowDist=dist;state.flowDirty=false}
 
-function playerMove(dx,dy,useDashKey=false){if(state.won||state.lost)return;const px=state.player.x,py=state.player.y;let didDash=false;if((state.dashArmed||useDashKey)&&state.dashCD===0&&state.mana>=DASH_COST){let nx=state.player.x,ny=state.player.y;for(let step=0;step<2;step++){const tx=nx+dx,ty=ny+dy;if(!inBounds(tx,ty)||isWall(tx,ty))break;nx=tx;ny=ty}if(nx!==state.player.x||ny!==state.player.y){state.player.x=nx;state.player.y=ny;state.mana-=DASH_COST;state.dashCD=DASH_CD;state.dashArmed=false;didDash=true;logMsg('Dashed!')}}if(!didDash){const nx=state.player.x+dx,ny=state.player.y+dy;if(!inBounds(nx,ny)||isWall(nx,ny))return;state.player.x=nx;state.player.y=ny}const mdx=state.player.x-px,mdy=state.player.y-py;if(mdx||mdy)state.lastMove={dx:mdx,dy:mdy};for(const c of state.map.chests){if(!c.opened&&samePos(c,state.player)){c.opened=true;state.mana+=CHEST_MANA;terrainValid=false;logMsg(`Opened chest: +${CHEST_MANA} mana.`)}}if(state.player.x===state.map.exit.x&&state.player.y===state.map.exit.y){state.won=true;updateHUD();return}state.flowDirty=true;advanceTurn()}
+function playerMove(dx,dy,useDashKey=false){if(state.won||state.lost)return;const px=state.player.x,py=state.player.y;let didDash=false;if((state.dashArmed||useDashKey)&&state.dashCD===0&&state.mana>=DASH_COST){let nx=state.player.x,ny=state.player.y;for(let step=0;step<2;step++){const tx=nx+dx,ty=ny+dy;if(!inBounds(tx,ty)||isWall(tx,ty))break;nx=tx;ny=ty}if(nx!==state.player.x||ny!==state.player.y){state.player.x=nx;state.player.y=ny;state.mana-=DASH_COST;state.dashCD=DASH_CD;state.dashArmed=false;didDash=true;logMsg('Dashed!')}}if(!didDash){const nx=state.player.x+dx,ny=state.player.y+dy;if(!inBounds(nx,ny)||isWall(nx,ny))return;state.player.x=nx;state.player.y=ny}const mdx=state.player.x-px,mdy=state.player.y-py;if(mdx||mdy)state.lastMove={dx:mdx,dy:mdy};for(const c of state.map.chests){if(!c.opened&&samePos(c,state.player)){c.opened=true;state.mana+=CHEST_MANA;terrainValid=false;logMsg(`Opened chest: +${CHEST_MANA} mana.`)}}updateCamera();if(state.player.x===state.map.exit.x&&state.player.y===state.map.exit.y){state.won=true;updateHUD();return}state.flowDirty=true;advanceTurn()}
 window.onMove=(dir)=>{if(dir==='up')playerMove(0,-1);else if(dir==='down')playerMove(0,1);else if(dir==='left')playerMove(-1,0);else if(dir==='right')playerMove(1,0);};
 function tryPlace(x,y){const t=state.selectedTool,cost=COSTS[t];state.mana-=cost;if(state.ammo[t]!==Infinity)state.ammo[t]--;if(t==='arrow')state.towers.push({x,y,type:t,ammo:ARROW_AMMO});else state.towers.push({x,y,type:t});state.placeMode=false;state.hover=null;logMsg(`Placed ${t} at (${x},${y}).`);updateMana(state.mana);advanceTurn()}
 function isValidPlacement(x,y){if(!inBounds(x,y))return{ok:false,reason:'out of bounds'};if(isWall(x,y))return{ok:false,reason:'wall tile'};if(isStart(x,y)||isExit(x,y))return{ok:false,reason:'reserved tile'};if(isSpawner(x,y))return{ok:false,reason:'spawner tile'};if(isChest(x,y))return{ok:false,reason:'chest tile'};if(state.player.x===x&&state.player.y===y)return{ok:false,reason:'on player'};const dist=Math.abs(state.player.x-x)+Math.abs(state.player.y-y);if(dist>PLACE_RADIUS)return{ok:false,reason:`must place within ${PLACE_RADIUS} tiles of player`};if(state.towers.some(t=>t.x===x&&t.y===y))return{ok:false,reason:'occupied by a trap'};const cost=COSTS[state.selectedTool];if(state.mana<cost)return{ok:false,reason:`need ${cost} mana`};return{ok:true}}
@@ -357,7 +352,7 @@ function handleSpawns(){
     if(state.turn>0&&state.turn%20===0)spawnBoss();
 }
 function checkWinLose(){if(state.player.x===state.map.exit.x&&state.player.y===state.map.exit.y){state.won=true;logMsg('You reached the exit. Victory!')}if(state.hp<=0){state.lost=true;logMsg('You have fallen...')}}
-function resetState(){const map=buildMap();state={map,turn:0,hp:START_HP,mana:START_MANA,nextSpawn:1,player:{x:map.start.x,y:map.start.y},enemies:[],towers:[],visited:Array.from({length:GRID_H},()=>Array(GRID_W).fill(false)),placeMode:false,selectedTool:'arrow',won:false,lost:false,fx:[],hover:null,flowDist:null,flowDirty:true,dashCD:0,dashArmed:false,lastMove:{dx:0,dy:0},ammo:{arrow:Infinity,rune:Infinity,fire:Infinity,spike:Infinity}};state.visited[state.player.y][state.player.x]=true;clearLog();logMsg('v2.9.7: trap icons, ammo meters, and fire totem AoE indicator.');const rect=canvas.getBoundingClientRect();const dpr=window.devicePixelRatio||1;canvas.width=Math.floor(rect.width*dpr);canvas.height=Math.floor(rect.height*dpr);ctx.setTransform(dpr,0,0,dpr,0,0);tileSize=rect.width/GRID_W;tilePad=Math.max(1,Math.floor(tileSize*.03));ensureOffscreen();terrainValid=false;renderTrapbar(TRAP_DEFS,state);TRAP_DEFS.forEach(t=>setCooldown(t.id,0,1));updateHUD();}
+function resetState(){const map=buildMap();state={map,turn:0,hp:START_HP,mana:START_MANA,nextSpawn:1,player:{x:map.start.x,y:map.start.y},enemies:[],towers:[],visited:Array.from({length:GRID_H},()=>Array(GRID_W).fill(false)),placeMode:false,selectedTool:'arrow',won:false,lost:false,fx:[],hover:null,flowDist:null,flowDirty:true,dashCD:0,dashArmed:false,lastMove:{dx:0,dy:0},ammo:{arrow:Infinity,rune:Infinity,fire:Infinity,spike:Infinity},cameraX:0,cameraY:0};state.visited[state.player.y][state.player.x]=true;clearLog();logMsg('v2.9.7: trap icons, ammo meters, and fire totem AoE indicator.');const rect=canvas.getBoundingClientRect();const dpr=window.devicePixelRatio||1;canvas.width=Math.floor(rect.width*dpr);canvas.height=Math.floor(rect.height*dpr);ctx.setTransform(dpr,0,0,dpr,0,0);tileSize=rect.width/VIEW_W;tilePad=Math.max(1,Math.floor(tileSize*.03));ensureOffscreen();terrainValid=false;updateCamera();renderTrapbar(TRAP_DEFS,state);TRAP_DEFS.forEach(t=>setCooldown(t.id,0,1));updateHUD();}
